refactor: [move-page] 사이드 네비게이션 바 코드 리팩토링
네비게이션 버튼 컴포넌트 분리
타입 분리해서 작성
li 내부 컴포넌트가 아닌 li에 key 속성 지정
