+++ FE/src/types/navigation.ts
@@ -1,0 +1,5 @@
+export interface NavigationInformation {
+  pageName: string;
+  description?: string;
+  pageURI: string;
+}
